Move readFromat* functions to separate files
